+++ apps/explorer/lib/explorer/chain/address.ex
@@ -19,7 +19,6 @@ defmodule Explorer.Chain.Address do
     Hash,
     InternalTransaction,
     SmartContract,
-    SmartContractAdditionalSource,
     Token,
     Transaction,
     Wei,
@@ -77,8 +76,7 @@ defmodule Explorer.Chain.Address do
              :token,
              :contracts_creation_internal_transaction,
              :contracts_creation_transaction,
-             :names,
-             :smart_contract_additional_sources
+             :names
            ]}
 
   @derive {Jason.Encoder,
@@ -89,8 +87,7 @@ defmodule Explorer.Chain.Address do
              :token,
              :contracts_creation_internal_transaction,
              :contracts_creation_transaction,
-             :names,
-             :smart_contract_additional_sources
+             :names
            ]}
 
   @primary_key {:hash, Hash.Address, autogenerate: false}
@@ -125,7 +122,6 @@ defmodule Explorer.Chain.Address do
 
     has_many(:names, Address.Name, foreign_key: :address_hash)
     has_many(:decompiled_smart_contracts, DecompiledSmartContract, foreign_key: :address_hash)
-    has_many(:smart_contract_additional_sources, SmartContractAdditionalSource, foreign_key: :address_hash)
     has_many(:withdrawals, Withdrawal, foreign_key: :address_hash)
 
     timestamps()