+++ apps/explorer/lib/explorer/chain/smart_contract_additional_sources.ex
@@ -15,11 +15,13 @@ defmodule Explorer.Chain.SmartContractAdditionalSource do
   @typedoc """
   * `file_name` - the name of the Solidity file with contract code (with extension).
   * `contract_source_code` - the Solidity source code from the file with `file_name`.
+  * `address_hash` - foreign key for `smart_contract`.
   """
 
   @type t :: %Explorer.Chain.SmartContractAdditionalSource{
           file_name: String.t(),
-          contract_source_code: String.t()
+          contract_source_code: String.t(),
+          address_hash: Hash.Address.t()
         }
 
   schema "smart_contracts_additional_sources" do